2-DataWarehouse-Transform/requirements.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
python-dotenv